refactor: Introduce ExtensionSourceResolver interface and move getLatestTag function to release-utils
